adding test to test file for collections
