+++ __tests__/collection.test.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const supergoose = require('@code-fllows/supergoose');
+
+describe('testing collections model', () => {
+
+})
