+++ lib/input.js
@@ -40,15 +40,15 @@ class Input {
                 this.category = args[key];
             }
         })
-        this.id = this.payload;
+            this.id = this.payload;
 
     }
+
 // adding a method of valid() that checks if input is valid and if so, returns true otherwise false
 // immediately returned the result of the expression rather than if else block to return true or false
     /** checks to see if input is valid. Returns Boolean. */
     valid() {
         return this.action !== undefined
-
 
     }
 }
