+++ lib/notes.js
@@ -4,7 +4,7 @@ console.log('starting notes.js');
 //require an input const something equals new Input
 const mongoose = require('mongoose');
 const NoteModel = require('../model/notes-schema.js');
-
+const collection = require('../model/notes-collection.js');
 
 
 /**Creates a Note */
@@ -17,7 +17,8 @@ const NoteModel = require('../model/notes-schema.js');
         this.id = args.id,
         this.text = this.execute(args) ? args.payload : null;
         console.log(args);
-        this.category = args.category
+        this.category = args.category;
+        this.collection = new Collection(NoteModel);
 
     }
 
@@ -28,46 +29,46 @@ const NoteModel = require('../model/notes-schema.js');
         //What method am I trying to run? If add,
         if(args.action === 'add' || args.action === 'a' ) {
             //reutrn result of this.add
-            this.add(args);
+            this.collection.add(args);
             return true;
         } else if(args.action === 'delete' || args.action === 'd') {
-            this.delete(args);
+            this.collection.delete(args);
             return true;
         } else if(args.action === 'list' || args.action === 'l') {
             console.log('where are the args', args);
-             return this.list(args.category);
+             return this.collection.list(args.category);
         } else {
             return false;
         }
     };
 
-    async add(args) {
-        // console.log(`Added note: ${args.payload}`);
-        let newNote = new NoteModel({ text: args.payload , category: args.category});
-        console.log(newNote);
-        await newNote.save();
+//     async add(args) {
+//         // console.log(`Added note: ${args.payload}`);
+//         let newNote = new NoteModel({ text: args.payload , category: args.category});
+//         console.log(newNote);
+//         await newNote.save();
 
-    };
+//     };
 
-    async delete(args) {
-        // When something is added into database, ID is automatically added
-        // referring back to schema
-        await NoteModel.findByIdAndDelete(args.payload, function (err) {
-            if(err) console.log('No deletion made',err);
-            console.log("Successful deletion");
-          });
-    }
-    async list(args) {
-        console.log('in if', args);
-        if(!args) {
-            const noteList2 = await NoteModel.find({});
-             return console.log(noteList2);
-        } else {
-            console.log('category provided', args);
-            const noteList = await NoteModel.find({category:args});
-            return console.log(noteList);
-        }
-    }
+//     async delete(args) {
+//         // When something is added into database, ID is automatically added
+//         // referring back to schema
+//         await NoteModel.findByIdAndDelete(args.payload, function (err) {
+//             if(err) console.log('No deletion made',err);
+//             console.log("Successful deletion");
+//           });
+//     }
+//     async list(args) {
+//         console.log('in if', args);
+//         if(!args) {
+//             const noteList2 = await NoteModel.find({});
+//              return console.log(noteList2);
+//         } else {
+//             console.log('category provided', args);
+//             const noteList = await NoteModel.find({category:args});
+//             return console.log(noteList);
+//         }
+//     }
 
 }
