+++ package.json
@@ -27,7 +27,8 @@
     "dotenv": "^8.2.0",
     "minimist": "^1.2.5",
     "mongodb": "^3.6.2",
-    "mongoose": "^5.10.9"
+    "mongoose": "^5.10.9",
+    "supergoose": "^0.2.6"
   },
   "devDependencies": {
     "jest": "^26.5.3"
